Update tasks tests to check balance updates after create_task, update_task and accept_task. Use u128 for AcccountId associated type in system::Config otherwise get_sub_account() can not produce unique value.

+++ pallets/task/src/mock.rs
@@ -44,7 +44,7 @@ impl system::Config for Test {
 	type BlockNumber = u64;
 	type Hash = H256;
 	type Hashing = BlakeTwo256;
-	type AccountId = u64;
+	type AccountId = u128;
 	type Lookup = IdentityLookup<Self::AccountId>;
 	type Header = Header;
 	type Event = Event;
